pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-db-repair-env"
version = "0.1.0"
description = "DB Repair Environment for OpenEnv - train AI agents to fix corrupted SQLite databases via SQL queries"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "db_repair_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["db_repair_env", "db_repair_env.server"]
package-dir = { "db_repair_env" = ".", "db_repair_env.server" = "server" }
